# backend/internal/dto/model_error_response.go
/*
 * Ban Do 5 Tot API
 *
 * Backend API for Ban Do 5 Tot project
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package dto

type ErrorResponse struct {
	Success bool `json:"success,omitempty"`

	Error ErrorResponseError `json:"error,omitempty"`
}