# backend/internal/dto/model_review_evidence_request.go
/*
 * Ban Do 5 Tot API
 *
 * Backend API for Ban Do 5 Tot project
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package dto

type ReviewEvidenceRequest struct {
	Status string `json:"status"`

	ReviewNote *string `json:"reviewNote,omitempty"`

	// Allow re-reviewing already reviewed evidence
	ForceOverride bool `json:"forceOverride,omitempty"`
}