# backend/internal/dto/model_update_activity_request.go
/*
 * Ban Do 5 Tot API
 *
 * Backend API for Ban Do 5 Tot project
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package dto

type UpdateActivityRequest struct {
	Title string `json:"title,omitempty"`

	Description *string `json:"description,omitempty"`

	Slug string `json:"slug,omitempty" validate:"regexp=^[a-z0-9-]+$"`

	ThumbnailUrl *string `json:"thumbnailUrl,omitempty"`

	ShortDescription *string `json:"shortDescription,omitempty"`

	UnitId *string `json:"unitId,omitempty"`

	StartDate *string `json:"startDate,omitempty"`

	EndDate *string `json:"endDate,omitempty"`

	IsActive bool `json:"isActive,omitempty"`

	RegistrationUrl *string `json:"registrationUrl,omitempty"`

	ReviewLevel *string `json:"reviewLevel,omitempty"`

	Organizer *string `json:"organizer,omitempty"`
}