# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Ban Do 5 Tot API
  description: Backend API for Ban Do 5 Tot project
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  responses:
    BadRequest:
      description: "Bad Request. The request is invalid."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Invalid request."
    Unauthorized:
      description: "Unauthorized. The provided JWT token is missing, invalid, or expired."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Authentication token is required."
    Forbidden:
      description: "Forbidden - You don't have permission to access this resource"
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Forbidden"
    NotFound:
      description: "Not Found. The requested resource was not found."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found."
    InternalServerError:
      description: "Internal Server Error. The server encountered an unexpected condition."
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Internal server error."
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    CriteriaDoc:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        maxScore:
          type: integer
    ActivityDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        thumbnailUrl:
          type: string
          format: uri
        unitId:
          type: string
          format: uuid
          nullable: true
        unitName:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        registrationUrl:
          type: string
          format: uri
          nullable: true
        reviewLevel:
          type: string
          enum: [TRUONG, DHQGHN, THANH_PHO, TRUNG_UONG]
        organizer:
          type: string
          nullable: true
        criteriaDocs:
          type: array
          items:
            $ref: '#/components/schemas/CriteriaDoc'
    EvidenceItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        activityId:
          type: string
          format: uuid
        activityTitle:
          type: string
        criteriaDocId:
          type: string
          format: uuid
          nullable: true
        criteriaDocTitle:
          type: string
          nullable: true
        fileUrl:
          type: string
          format: uri
        description:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewNote:
          type: string
          nullable: true
        reviewedBy:
          type: string
          format: uuid
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    ReviewEvidenceRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [approved, rejected]
        reviewNote:
          type: string
          nullable: true
        forceOverride:
          type: boolean
          default: false
          description: Allow re-reviewing already reviewed evidence
    CreateActivityRequest:
      type: object
      required:
        - title
        - slug
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
          description: URL-friendly slug, must be unique
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        shortDescription:
          type: string
          nullable: true
        unitId:
          type: string
          format: uuid
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        isActive:
          type: boolean
          default: true
        registrationUrl:
          type: string
          format: uri
          nullable: true
        reviewLevel:
          type: string
          enum: [TRUONG, DHQGHN, THANH_PHO, TRUNG_UONG]
          nullable: true
        organizer:
          type: string
          nullable: true
    UpdateActivityRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
        shortDescription:
          type: string
          nullable: true
        unitId:
          type: string
          format: uuid
          nullable: true
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
        isActive:
          type: boolean
        registrationUrl:
          type: string
          format: uri
          nullable: true
        reviewLevel:
          type: string
          enum: [TRUONG, DHQGHN, THANH_PHO, TRUNG_UONG]
          nullable: true
        organizer:
          type: string
          nullable: true
paths:
...
  /admin/evidences/{id}/review:
    patch:
      summary: Review an evidence submission
      operationId: reviewEvidence
      tags:
        - Admin
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Evidence ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewEvidenceRequest'
      responses:
        '200':
          description: Evidence reviewed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/EvidenceItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Evidence already reviewed and forceOverride not set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /admin/activities:
    post:
      summary: Create a new activity
      operationId: createActivity
      tags:
        - Admin
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Activity created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ActivityDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Activity with this slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/admin/activities/{id}':
    patch:
      summary: Update an activity
      operationId: updateActivity
      tags:
        - Admin
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Activity ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActivityRequest'
      responses:
        '200':
          description: Activity updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ActivityDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete an activity
      operationId: deleteActivity
      tags:
        - Admin
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Activity ID
      responses:
        '200':
          description: Activity deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'